backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
aiohttp==3.11.10
aiosqlite==0.20.0
pydantic==2.10.4
python-dotenv==1.0.1
websockets==14.1
python-dateutil==2.9.0
